Job-Fit/requirements.txt
sentence-transformers
pdfminer.six
numpy
scikit-learn
streamlit
pandas
streamlit-lottie
